refactor(js/*): migrate to ContextMenu component
