+++ src/dev.html
@@ -18,26 +18,7 @@
   </head>
 
   <body>
-    <ul id="contextMenus" class="dropdown-menu" role="menu" style="display:none">
-      <li id="cmenu_copy" class="cmenuItem contextSel"><a></a></li>
-      <li id="cmenu_copyAnsi" class="cmenuItem contextSel"><a></a></li>
-      <li id="cmenu_paste" class="cmenuItem contextNormal"><a></a></li>
-      <li id="cmenu_searchGoogle" class="cmenuItem contextSel"><a></a></li>
-      <li id="cmenu_openUrlNewTab" class="cmenuItem contextUrl"><a></a></li>
-      <li id="cmenu_copyLinkUrl" class="cmenuItem contextUrl"><a></a></li>
-      <li id="cmenu_divider1" class="divider"></li>
-      <li id="cmenu_quickSearch" class="cmenuItem contextQuickSearch"><a></a>
-        <ul id="quickSearchMenus" class="dropdown-menu" role="menu">
-        </ul>
-      </li>
-      <li id="cmenu_divider2" class="divider contextQuickSearch"></li>
-      <li id="cmenu_selectAll" class="cmenuItem contextNormal"><a></a></li>
-      <li id="cmenu_mouseBrowsing" class="cmenuItem contextNormal"><a></a></li>
-      <li id="cmenu_showInputHelper" class="cmenuItem contextNormal"><a></a></li>
-      <li id="cmenu_showLiveArticleHelper" class="cmenuItem contextNormal"><a></a></li>
-      <li id="cmenu_divider3" class="divider contextNormal"></li>
-      <li id="cmenu_settings" class="cmenuItem"><a></a></li>
-    </ul>
+    <div id="cmenuReact"></div>
 
     <div id="BBSWindow" style="display:none">
       <input id="t" type="text" autocomplete="off" style="box-sizing:content-box; color:initial; border: none; width: 200px; height: 24px; position:absolute; left:-10000px; top:0px; z-index:2;resize: none; font-family:inherit !important;" autofocus/>
