+++ src/js/main.js
@@ -1,7 +1,7 @@
 ﻿import { App } from './pttchrome';
 import { setupI18n } from './i18n';
 import { getQueryVariable } from './util';
-import { readValuesWithDefault } from '../components/PrefModal';
+import { readValuesWithDefault } from '../components/ContextMenu/PrefModal';
 
 function startApp() {
   var site = getQueryVariable('site');
