+++ src/js/pttchrome.js
@@ -13,9 +13,7 @@ import { unescapeStr, b2u, parseWaterball } from './string_util';
 import { getQueryVariable, setTimer } from './util';
 import PasteShortcutAlert from '../components/PasteShortcutAlert';
 import ConnectionAlert from '../components/ConnectionAlert';
-import InputHelperModal from '../components/InputHelperModal';
-import LiveHelperModal from '../components/LiveHelperModal';
-import PrefModal from '../components/PrefModal';
+import ContextMenu from '../components/ContextMenu';
 
 function noop() {}
 
@@ -1270,379 +1268,13 @@ App.prototype.mouse_scroll = function(e) {
   }
 };
 
-App.prototype.showQuickSearchMenus = function(e, selectedText, hideContextMenu) {
-  var self = this;
-  if (QUICK_SEARCH.providers.length === 0) return;
-
-  var menuSelector = '#quickSearchMenus';
-  var menuHtml = '';
-  for (var i = 0; i < QUICK_SEARCH.providers.length; ++i) {
-    var q = QUICK_SEARCH.providers[i];
-    menuHtml += '<li class="cmenuItem"><a data-url="'+q.url+'">'+q.name+'</a></li>';
-  }
-  $(menuSelector).html(menuHtml);
-  
-  $('#quickSearchMenus a').off();
-  $('#quickSearchMenus a').click(function(e) {
-    var url = $(this).data('url');
-    url = url.replace('%s', selectedText);
-    window.open(url);
-    e.stopPropagation();
-    hideContextMenu();
-  });
-
-  var pageHeight = $(window).height();
-  var pageWidth = $(window).width();
-  if (e.pageY > pageHeight/2) {
-    $(menuSelector).addClass('cmenuGoesUp');
-  } else {
-    $(menuSelector).removeClass('cmenuGoesUp');
-  }
-  if (e.pageX > pageWidth * 0.8) {
-    $(menuSelector).addClass('cmenuGoesLeft');
-  } else {
-    $(menuSelector).removeClass('cmenuGoesLeft');
-  }
-};
-
 App.prototype.setupContextMenus = function() {
-  var self = this;
-  var menuSelector = '#contextMenus';
-  var selectedText = '';
-  var contextOnUrl = '';
-  var contextOnUserId = '';
-  var aElement = null;
-
-  $('#BBSWindow').on('contextmenu', function(e) {
-    // if i am doing scrolling, i should skip
-    var cmdhandler = self.CmdHandler;
-    var doDOMMouseScroll = (cmdhandler.getAttribute('doDOMMouseScroll')=='1');
-    if (doDOMMouseScroll) {
-      e.stopPropagation();
-      e.preventDefault();
-      cmdhandler.setAttribute('doDOMMouseScroll','0');
-      return;
-    }
-
-    var target = $(e.target);
-    contextOnUrl = '';
-    contextOnUserId = '';
-
-    // just in case the selection get de-selected
-    if (window.getSelection().isCollapsed) {
-      self.lastSelection = null;
-    } else {
-      self.lastSelection = self.view.getSelectionColRow();
-    }
-
-    aElement = null;
-    if (target.is('a')) {
-      contextOnUrl = target.attr('href');
-      aElement = target[0];
-    } else if (target.parent().is('a')) {
-      contextOnUrl = target.parent().attr('href');
-      aElement = target[0].parentNode;
-    }
-
-    // replace the &nbsp;
-    selectedText = window.getSelection().toString().replace(/\u00a0/g, " ");
-
-    if (contextOnUrl) {
-      $('.contextQuickSearch').hide();
-      $('.contextUrl').show();
-      $('.contextSel').hide();
-      $('.contextNormal').hide();
-    } else {
-      if (window.getSelection().isCollapsed) { 
-        $('.contextQuickSearch').hide();
-        $('.contextUrl').hide();
-        $('.contextSel').hide();
-        $('.contextNormal').show();
-      } else {
-        // got something selected, show copy and searchGoogle
-        $('.contextUrl').hide();
-        $('.contextSel').show();
-        $('.contextNormal').hide();
-        var clipedText = selectedText;
-        if (clipedText.length > 15) {
-          clipedText = clipedText.substr(0, 15) + ' ... ';
-        }
-        $('#cmenuSearchContent').text("'"+clipedText+"'");
-        if (QUICK_SEARCH.providers.length > 0) {
-          self.showQuickSearchMenus(e, selectedText, function() {
-            hideContextMenu();
-          });
-          $('.contextQuickSearch').show();
-        } else {
-          $('.contextQuickSearch').hide();
-        }
-      }
-    }
-
-    // check if mouse browsing is on
-    if (self.buf.useMouseBrowsing) {
-      $('#cmenu_mouseBrowsing a').addClass('checked');
-    } else {
-      $('#cmenu_mouseBrowsing a').removeClass('checked');
-    }
-
-    // show and position
-    $(menuSelector)
-      .show()
-      .css({
-        position: "absolute",
-        left: function(e) {
-          var mouseWidth = e.pageX;
-          var pageWidth = $(window).width();
-          var menuWidth = $(menuSelector).width();
-          
-          // opening menu would pass the side of the page
-          if (mouseWidth + menuWidth > pageWidth &&
-              menuWidth < mouseWidth) {
-              return mouseWidth - menuWidth;
-          } 
-          return mouseWidth;
-        }(e),
-        top: function(e) {
-          var mouseHeight = e.pageY;
-          var pageHeight = $(window).height();
-          var menuHeight = $(menuSelector).height();
-
-          // opening menu would pass the bottom of the page
-          if (mouseHeight + menuHeight > pageHeight &&
-              menuHeight < mouseHeight) {
-              return mouseHeight - menuHeight;
-          } 
-          return mouseHeight;
-        }(e)
-      });
-    self.contextMenuShown = true;
-    return false;
-  });
-
-  var hideContextMenu = function() {
-    $(menuSelector).hide();
-    selectedText = '';
-    self.contextMenuShown = false;
-  };
-
-  // for menuitem shortkey
-  $(window).keyup(function(e) {
-    if (!self.contextMenuShown)
-      return;
-    if (!e.altKey && !e.ctrlKey && !e.shiftKey) {
-      switch (e.keyCode) {
-        case 67: // C
-          self.doCopy(selectedText);
-          hideContextMenu();
-          break;
-        case 69: // E
-          self.doCopy(contextOnUrl);
-          hideContextMenu();
-          break;
-        case 80: // P
-          self.doPaste();
-          hideContextMenu();
-          break;
-        case 83: // S
-          self.doSearchGoogle(selectedText);
-          hideContextMenu();
-          break;
-        case 84: // T
-          self.doOpenUrlNewTab(aElement);
-          hideContextMenu();
-          break;
-      }
-    }
-    e.preventDefault();
-    e.stopPropagation();
-  });
-
-  //make sure menu closes on any click
-  $(window).click(function() {
-    hideContextMenu();
-  });
-  if ('ontouchstart' in window) {
-    window.ontouchstart = function(e) {
-      if (e.target.parentNode.classList.length > 0 &&
-          e.target.parentNode.classList[0] == 'cmenuItem') {
-        return;
-      }
-      hideContextMenu();
-    };
-  }
-
-  $('#cmenu_copy a').html(i18n('cmenu_copy')+'<span class="cmenuHotkey">Ctrl+C</span>');
-  $('#cmenu_copyAnsi a').text(i18n('cmenu_copyAnsi'));
-  $('#cmenu_paste a').html(i18n('cmenu_paste')+'<span class="cmenuHotkey">Shift+Insert</span>');
-  $('#cmenu_selectAll a').html(i18n('cmenu_selectAll')+'<span class="cmenuHotkey">Ctrl+A</span>');
-  $('#cmenu_searchGoogle a').html(i18n('cmenu_searchGoogle')+' <span id="cmenuSearchContent"></span>');
-  $('#cmenu_quickSearch a').html(i18n('cmenu_quickSearch')+' <span style="float:right;">&#9658;</span>');
-  $('#cmenu_openUrlNewTab a').text(i18n('cmenu_openUrlNewTab'));
-  $('#cmenu_copyLinkUrl a').text(i18n('cmenu_copyLinkUrl'));
-  $('#cmenu_mouseBrowsing a').text(i18n('cmenu_mouseBrowsing'));
-  $('#cmenu_showInputHelper a').text(i18n('cmenu_showInputHelper'));
-  $('#cmenu_showLiveArticleHelper a').text(i18n('cmenu_showLiveArticleHelper'));
-  $('#cmenu_settings a').text(i18n('cmenu_settings'));
-
-  var contextMenuItemOnClickHandler = {
-    'cmenu_copy': function() { 
-      self.doCopy(selectedText); 
-    },
-    'cmenu_copyAnsi': function() { 
-      self.doCopyAnsi(); 
-    },
-    'cmenu_paste': function() { 
-      self.doPaste(); 
-    },
-    'cmenu_selectAll': function() { 
-      self.doSelectAll(); 
-    },
-    'cmenu_searchGoogle': function() { 
-      self.doSearchGoogle(selectedText); 
-    },
-    'cmenu_openUrlNewTab': function() { 
-      self.doOpenUrlNewTab(aElement); 
-    },
-    'cmenu_copyLinkUrl': function() {
-      self.doCopy(contextOnUrl); 
-    },
-    'cmenu_mouseBrowsing': function() { 
-      self.switchMouseBrowsing(); 
-    },
-    'cmenu_showInputHelper': () => {
-      const onReset = () => this.conn.send('\x15[m');
-      const onCmdSend = cmd => {
-        if (!window.getSelection().isCollapsed && this.buf.pageState == 6) {
-          // something selected
-          var sel = this.view.getSelectionColRow();
-          var y = this.buf.cur_y;
-          var selCmd = '';
-          // move cursor to end and send reset code
-          selCmd += '\x1b[H';
-          if (y > sel.end.row) {
-            selCmd += '\x1b[A'.repeat(y - sel.end.row);
-          } else if (y < sel.end.row) {
-            selCmd += '\x1b[B'.repeat(sel.end.row - y);
-          }
-          var repeats = this.buf.getRowText(sel.end.row, 0, sel.end.col).length;
-          selCmd += '\x1b[C'.repeat(repeats) + '\x15[m';
-      
-          // move cursor to start and send color code
-          y = sel.end.row;
-          selCmd += '\x1b[H';
-          if (y > sel.start.row) {
-            selCmd += '\x1b[A'.repeat(y - sel.start.row);
-          } else if (y < sel.start.row) {
-            selCmd += '\x1b[B'.repeat(sel.start.row - y);
-          }
-          repeats = this.buf.getRowText(sel.start.row, 0, sel.start.col).length;
-          selCmd += '\x1b[C'.repeat(repeats);
-          cmd = selCmd + cmd;
-        }
-        this.conn.send(cmd);
-      }
-      const onConvSend = value => this.conn.convSend(value);
-      const onHide = () => {
-        ReactDOM.unmountComponentAtNode(document.getElementById('reactAlert'))
-      }
-      ReactDOM.render(
-        <InputHelperModal
-          onHide={onHide}
-          onReset={onReset}
-          onCmdSend={onCmdSend}
-          onConvSend={onConvSend}
-        />,
-        document.getElementById('reactAlert')
-      );
-    },
-    'cmenu_showLiveArticleHelper': () => {
-      let state = {
-        enabled: false,
-        sec: 1,
-      };
-      const onHide = () => {
-        this.onToggleLiveHelperModalState = noop;
-        this.onDisableLiveHelperModalState = noop;
-        ReactDOM.unmountComponentAtNode(document.getElementById('reactAlert'))
-      };
-      const onChange = nextState => {
-        state = nextState;
-        renderLiveHelperModal();
-        if (state.enabled) {
-          // cancel easy reading mode first
-          this.view.useEasyReadingMode = false;
-          this.switchToEasyReadingMode();
-          this.setAutoPushthreadUpdate(state.sec);
-        } else {
-          this.setAutoPushthreadUpdate(-1);
-        }
-      };
-      const renderLiveHelperModal = () => {
-        ReactDOM.render(
-          <LiveHelperModal
-            onHide={onHide}
-            {...state}
-            onChange={onChange}
-          />,
-          document.getElementById('reactAlert')
-        );
-      };
-      renderLiveHelperModal();
-      //
-      this.onToggleLiveHelperModalState = () => {
-        onChange({
-          enabled: !state.enabled,
-          sec: state.sec,
-        })
-      };
-      this.onDisableLiveHelperModalState = () => {
-        onChange({
-          enabled: false,
-          sec: state.sec,
-        })
-      };
-    },
-    'cmenu_settings': () => { 
-      const onSave = (values) => {
-        ReactDOM.unmountComponentAtNode(document.getElementById('reactAlert'))
-        this.onValuesPrefChange(values);
-        this.modalShown = false;
-        this.setInputAreaFocus();
-        this.switchToEasyReadingMode(this.view.useEasyReadingMode);
-      };
-      const onReset = (values) => {
-        this.view.redraw(true);
-        onSave(values)
-      }
-      const renderPrefModal = () => {
-        this.onDisableLiveHelperModalState();
-        this.modalShown = true;
-        ReactDOM.render(
-          <PrefModal
-            onSave={onSave}
-            onReset={onReset}
-          />,
-          document.getElementById('reactAlert')
-        );
-      };
-      renderPrefModal();
-    }
-  };
-
-  $('.cmenuItem').click(function(e) {
-    var id = $(this).attr('id');
-    if (id in contextMenuItemOnClickHandler) {
-      contextMenuItemOnClickHandler[id]();
-    }
-    e.stopPropagation();
-    hideContextMenu();
-  });
-
-  $(menuSelector).on('contextmenu', function(e) {
-    e.stopPropagation();
-    e.preventDefault();
-  });
+  ReactDOM.render(
+    <ContextMenu
+      pttchrome={this}
+    />,
+    document.getElementById('cmenuReact')
+  );
 };
 
 App.prototype.context_menu = function(e) {
